refactor: modularize quote/[id]/page.tsx — 749→163 lines, 15 components, responsive
Extract 12 inline components into separate files with proper separation:
- lib/types.ts, format.ts, constants.ts, useMediaQuery.ts (shared utilities)
- components/ui: IconBtn, TabBtn, FileLink (reusable leaf components)
- components/quote: Section, MetaItem, ReqField, InfoBar, PricingSummary,
  SourceFilesSection, QuoteHeader, QuoteTabBar, DetailView, ChatView
- components/chat: ChatInput (extracted with drag-drop logic)

Adds responsive breakpoints (mobile ≤640px, tablet ≤1024px, desktop),
React.memo on 6 leaf components, useMemo on 3 breakdown calculations,
and resolves UIMessage circular import via lib/types.ts.
