+++ web/src/lib/types.ts
@@ -1,0 +1,7 @@
+export interface UIMessage {
+  id: string;
+  role: "user" | "assistant";
+  content: string;
+  isStreaming?: boolean;
+  attachmentName?: string;
+}
